poller: Correctly distinguish backlog from 'live' messages
I wrote this code before understanding what a state event is, so I
assumed the "state" key was for the state before we polled, and the
"timeline" key were the live changes; and based the rest of the code on
this.

This is obviously completely wrong, and caused a bunch of complex bugs
(though it worked most of the time).

This should fix most of them, and I'll deal with the rest as I encounter
them.