Do not send backlog events
It is messy to resend then when clients restart, because none of them
deduplicates messages on msgi.

Some might if we wrapped the backlog in a 'chathistory' batch, but 1.
not all of them would, 2. it's hard to implement correctly.

Therefore, let's remove it entirely. If clients want the backlog, they
should implement support for the CHATHISTORY command.